Fixed various issues
