+++ src/Checksum.hpp
@@ -6,3 +6,4 @@
 bool AddDemoChecksum(const char *filename);
 void AddDemoFileChecksums();
 void AddDemoVpkChecksums();
+void RecordRuntimeVscriptChecksum(const char *scriptName, const char *scriptData);
